feat: Reservation 과 SignIn 파일 및 폴더 컨벤션 정리
