+++ app/components/screens/reservation/index.ts
@@ -1,2 +1,3 @@
-export * from './process';
-export * from './time-table';
+export * from './reservation-mentoring-process';
+export * from './reservation-band-process';
+export * from './reservation-time-table';
