+++ app/components/screens/reservation/reservation-band-process/index.ts
@@ -1,0 +1,1 @@
+export * from './reservation-band-process.screen';
