+++ app/components/screens/reservation/reservation-mentoring-process/index.ts
@@ -1,0 +1,1 @@
+export * from './reservation-mentoring-process.screen';
