+++ app/components/screens/reservation/reservation-mentoring-process/reservation-mentoring-process.screen.tsx
@@ -1,30 +1,26 @@
-import React, {useEffect, useState} from 'react';
+import React, {useState} from 'react';
 import {
   NavigationProp,
   ParamListBase,
   useNavigation,
 } from '@react-navigation/native';
 import DropDownPicker from 'react-native-dropdown-picker';
-import {View, StyleSheet, Text, Platform, ScrollView} from 'react-native';
-import CustomBtn from '../../../layout/cta-button.layout';
-import color from '../../../../utils/data/color/type';
+import {View, StyleSheet, Text, Platform} from 'react-native';
+import {ScreenWrapper, CTAButton, Modal, customBtnType} from '../../../layout';
 import {
   dayItems,
   PROCESS_TEXT,
-  RESERVATION_TYPE,
+  sectionItems,
   timeItems,
-  times,
-} from '../../../../utils/constant/reservation/process/reservationProcess';
-import CustomModal from '../../../layout/modal.layout';
+  unitItems,
+} from '../reservation.data';
 import {
   fontPercentage,
   heightPercentage,
   widthPercentage,
-} from '../../../../utils/api/responsive/responsive.api';
-import ScreenWrapper from '../../../layout/screen-wrapper.layout';
-import {customBtnType} from '../../../../utils/types/customModal';
-import {ItemType} from '../../../../utils/types/drop-down.type';
-import {postReservation} from '../../../../utils/api/axios';
+  ItemType,
+  ValueType,
+} from '../../../../utils';
 
 const styles = StyleSheet.create({
   root: {
@@ -62,18 +58,7 @@ const styles = StyleSheet.create({
       android: {},
     }),
   },
-  startTimePicker: {
-    width: '100%',
-    height: heightPercentage(46),
-    marginTop: heightPercentage(20),
-    ...Platform.select({
-      ios: {
-        zIndex: 90,
-      },
-      android: {},
-    }),
-  },
-  endTimePicker: {
+  reservationTimePicker: {
     width: '100%',
     height: heightPercentage(46),
     marginTop: heightPercentage(20),
@@ -186,32 +171,10 @@ const styles = StyleSheet.create({
     marginTop: heightPercentage(6),
     color: '#000000',
   },
-  scrollTime: {
-    fontFamily: 'NotoSansKR-Regular',
-    fontSize: fontPercentage(9),
-    color: '#6d6d6d',
-    alignItems: 'center',
-    justifyContent: 'center',
-    alignContent: 'center',
-  },
-  scrollTimeBox: {
-    width: widthPercentage(19),
-    height: heightPercentage(11),
-    borderStyle: 'solid',
-    borderWidth: 1,
-    borderColor: '#707070',
-    marginTop: heightPercentage(3),
-    marginLeft: widthPercentage(3),
-  },
-  alignCenter: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
 });
 
-function BandReservationProcess({route}: any) {
-  const {currentWeek, monday} = route.params;
-
+// eslint-disable-next-line react/prop-types
+export function ReservationMentoringProcess({route}) {
   const navigation: NavigationProp<ParamListBase> = useNavigation();
   const [modalVisible, setModalVisible] = useState<boolean>(false);
 
@@ -229,55 +192,29 @@ function BandReservationProcess({route}: any) {
     },
   ];
 
-  const [day, setDay] = useState<string>('');
+  // eslint-disable-next-line react/prop-types
+  const currentWeek: any = route.params;
+
+  const [day, setDay] = useState<ValueType>('');
   const [dayOpen, setDayOpen] = useState<boolean>(false);
   const [dayItem, setDayItems] = useState<Array<ItemType>>(dayItems);
 
-  const [startTime, setStartTime] = useState<number | null>(null);
-  const [startTimeOpen, setStartTimeOpen] = useState<boolean>(false);
-  const [startTimeItem, setStartTimeItems] =
-    useState<Array<ItemType>>(timeItems);
+  const [unit, setUnit] = useState<ValueType>('');
+  const [unitOpen, setUnitOpen] = useState<boolean>(false);
+  const [unitItem, setUnitItems] = useState<Array<ItemType>>(unitItems);
 
-  const [endTime, setEndTime] = useState<number | null>(null);
-  const [endTimeOpen, setEndTimeOpen] = useState<boolean>(false);
-  const [endTimeItem, setEndTimeItems] = useState<Array<ItemType>>(timeItems);
+  const [time, setTime] = useState<ValueType>('');
+  const [timeOpen, setTimeOpen] = useState<boolean>(false);
+  const [timeItem, setTimeItems] = useState<Array<ItemType>>(timeItems);
 
-  const submitBtnClickListener = () => {
-    if (startTime !== null && endTime !== null) {
-      postReservation(
-        monday,
-        RESERVATION_TYPE.Together,
-        day,
-        startTime,
-        endTime,
-        '',
-        '',
-      )
-        .then((res) => {
-          console.log(JSON.stringify(res, null, 2));
-          changeVisible();
-        })
-        .catch((err) => {
-          console.log(JSON.stringify(err, null, 2));
-        });
-    }
-  };
-
-  useEffect(() => {
-    if (startTime !== null) {
-      const index = timeItems.findIndex((i) => i.value === startTime);
-
-      setEndTimeItems(timeItems.slice(index + 1));
-    }
-  }, [startTime]);
-
-  useEffect(() => {
-    console.log(startTime);
-  }, [startTime]);
+  const [section, setSection] = useState<ValueType[]>([]);
+  const [sectionOpen, setSectionOpen] = useState<boolean>(false);
+  const [sectionItem, setSectionItems] =
+    useState<Array<ItemType>>(sectionItems);
 
   return (
-    <ScreenWrapper headerTitle="고정합주 예약하기">
-      <CustomModal
+    <ScreenWrapper headerTitle="멘토링 예약하기">
+      <Modal
         mdVisible={modalVisible}
         title={'예약이 완료되었습니다!'}
         buttonList={modalBtn}
@@ -305,63 +242,68 @@ function BandReservationProcess({route}: any) {
 
         <View style={styles.contentContainer}>
           <View style={styles.timeBox}>
-            <ScrollView horizontal={true}>
-              <View style={styles.alignCenter}>
-                {times.map((item) => {
-                  return (
-                    <>
-                      <View>
-                        <Text style={styles.scrollTime}>{item.label}</Text>
-                        <View style={styles.scrollTimeBox}></View>
-                      </View>
-                      <View>
-                        <Text style={styles.scrollTime}> </Text>
-                        <View style={styles.scrollTimeBox}></View>
-                      </View>
-                    </>
-                  );
-                })}
-              </View>
-            </ScrollView>
+            <Text> {`시간들이 들어갈 공간`}</Text>
           </View>
-          <View style={styles.startTimePicker}>
+          <View style={styles.UnitPicker}>
             <DropDownPicker
-              open={startTimeOpen}
-              value={startTime}
-              items={startTimeItem}
-              setOpen={setStartTimeOpen}
-              setValue={setStartTime}
-              setItems={setStartTimeItems}
+              open={unitOpen}
+              value={unit}
+              items={unitItem}
+              setOpen={setUnitOpen}
+              setValue={setUnit}
+              setItems={setUnitItems}
               style={styles.dropDown2}
               textStyle={styles.dropDownText}
               dropDownContainerStyle={styles.dropDownContainer}
               placeholderStyle={styles.dropDownPlaceHolder}
-              placeholder={PROCESS_TEXT.START}
+              placeholder={PROCESS_TEXT.UNIT}
               zIndex={9000}
             />
           </View>
-          <View style={styles.endTimePicker}>
+          <View style={styles.reservationTimePicker}>
             <DropDownPicker
-              open={endTimeOpen}
-              value={endTime}
-              items={endTimeItem}
-              setOpen={setEndTimeOpen}
-              setValue={setEndTime}
-              setItems={setEndTimeItems}
+              open={timeOpen}
+              value={time}
+              items={timeItem}
+              setOpen={setTimeOpen}
+              setValue={setTime}
+              setItems={setTimeItems}
               style={styles.dropDown2}
               textStyle={styles.dropDownText}
               dropDownContainerStyle={styles.dropDownContainer}
               placeholderStyle={styles.dropDownPlaceHolder}
-              placeholder={PROCESS_TEXT.END}
+              placeholder={PROCESS_TEXT.TIME}
               zIndex={8000}
             />
           </View>
+          <Text style={styles.sectionInfo__alert__text}>
+            {PROCESS_TEXT.ALERT}
+          </Text>
+          <View style={styles.sectionInfo__form}>
+            <DropDownPicker
+              multiple={true}
+              min={0}
+              max={3}
+              open={sectionOpen}
+              value={section}
+              items={sectionItem}
+              setOpen={setSectionOpen}
+              setValue={setSection}
+              setItems={setSectionItems}
+              style={styles.dropDown2}
+              textStyle={styles.dropDownText}
+              dropDownContainerStyle={styles.dropDownContainer}
+              placeholderStyle={styles.dropDownPlaceHolder}
+              placeholder={PROCESS_TEXT.SECTION}
+              zIndex={7000}
+            />
+          </View>
           <View style={styles.submit}>
-            <CustomBtn
+            <CTAButton
               title={PROCESS_TEXT.SUBMIT}
               btnStyle={styles.submit__btn}
               titleStyle={styles.submit__text}
-              onClickListener={submitBtnClickListener}
+              onClickListener={changeVisible}
             />
           </View>
         </View>
@@ -369,5 +311,3 @@ function BandReservationProcess({route}: any) {
     </ScreenWrapper>
   );
 }
-
-export default BandReservationProcess;
