+++ app/components/screens/reservation/reservation-time-table/components/index.ts
@@ -1,0 +1,1 @@
+export * from './time-table.reservation.component';
