+++ app/components/screens/reservation/reservation-time-table/index.ts
@@ -1,0 +1,1 @@
+export * from './reservation-time-table.screen';
