+++ app/components/screens/sign-in/components/index.ts
@@ -1,0 +1,1 @@
+export * from './form.sign-in.component';
